+++ invoicing/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class InvoicingConfig(AppConfig):
+    name = 'invoicing'
