+++ invoicing/views.py
@@ -4,6 +4,8 @@ from django_filters.views import object_filter
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 
 """This file contains the abstract views for Invoicing"""
+
+
 class InvoiceListView(ListView):
     model = None
     filterset_class = None
@@ -12,14 +14,16 @@ class InvoiceListView(ListView):
     def get_queryset(self):
 
         queryset = super().get_queryset()
-        self.filterset = self.filterset_class(self.request.GET, queryset=queryset)
+        self.filterset = self.filterset_class(
+            self.request.GET, queryset=queryset)
 
         return self.filterset.qs
-    
+
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
         context['filter'] = self.filterset
         return context
+
 
 class InvoiceCreateView(LoginRequiredMixin, CreateView):
     model = None
@@ -28,9 +32,9 @@ class InvoiceCreateView(LoginRequiredMixin, CreateView):
         form.instance.organization = self.request.user.organization
         return super().form_valid(form)
 
+
 class InvoiceUpdateView(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
     model = None
-
 
     def form_valid(self, form):
         form.instance.organization = self.request.user.organization
@@ -40,6 +44,7 @@ class InvoiceUpdateView(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
         invoice = self.get_object()
         # Don't let the user edit if invoice is finalized
         return invoice.organization == self.request.user.organization and not invoice.finalized
+
 
 class InvoiceDetailView(LoginRequiredMixin, UserPassesTestMixin, DetailView):
     model = None
@@ -54,6 +59,7 @@ class InvoiceDetailView(LoginRequiredMixin, UserPassesTestMixin, DetailView):
     def test_func(self):
         invoice = self.get_object()
         return invoice.organization == self.request.user.organization
+
 
 class InvoiceDeleteView(LoginRequiredMixin, UserPassesTestMixin, DeleteView):
     model = None
@@ -89,6 +95,7 @@ class InvoiceEntryCreateView(LoginRequiredMixin, UserPassesTestMixin, CreateView
             self.parent_model, pk=self.kwargs.get("pk"))
         # Don't let the user edit if invoice is finalized
         return self.invoice.organization == self.request.user.organization and not self.invoice.finalized
+
 
 class InvoiceEntryDeleteView(LoginRequiredMixin, UserPassesTestMixin, DeleteView):
     model = None
