+++ products/forms.py
@@ -11,14 +11,3 @@ class ProductForm(forms.ModelForm):
             'sale_price': 'MRP',
             'unit': 'Price Unit'
         }
-
-"""Abstract form for Invoicing"""
-class InvoiceForm(forms.ModelForm):
-
-    date = forms.DateField(widget=forms.TextInput(attrs={
-        'type': 'date',
-        'value': timezone.now().strftime("%Y-%m-%d")
-    }))
-
-    class Meta:
-        exclude = ('organization', 'finalized')
