+++ sales/filters.py
@@ -1,13 +1,7 @@
 from .models import SalesInvoice
-import django_filters
+from invoicing.filters import InvoiceFilter
 
 
-class SalesInvoiceFilter(django_filters.FilterSet):
-    name = django_filters.CharFilter(lookup_expr='icontains')
-    date = django_filters.DateFromToRangeFilter(widget=django_filters.widgets.RangeWidget(attrs={
-        'type': 'date'
-    }))
-
-    class Meta:
+class SalesInvoiceFilter(InvoiceFilter):
+    class Meta(InvoiceFilter.Meta):
         model = SalesInvoice
-        fields = ['name', 'date']
